requirements.txt
Django<1.9
django-extensions
django-flat-theme
django-localflavor
django-reversion
fake-factory
ipython
psycopg2
pytest-django
requests
